feat: adding super class for entity builds
