+++ plugins/dependency-packages-backend-module-npm-provider/src/builder/NpmComponentEntityBuilder.ts
@@ -1,38 +1,20 @@
 import { ComponentEntity, ANNOTATION_LOCATION, ANNOTATION_ORIGIN_LOCATION, ANNOTATION_SOURCE_LOCATION } from "@backstage/catalog-model";
-import { EntityTemplate, ParsedDependencies } from "@voodoogq/plugin-dependency-packages-common";
+import { ParsedDependencies } from "@voodoogq/plugin-dependency-packages-common";
 import { RegistryMetaRetriever } from "../retriever/RegistryMetaRetriever";
 import { NpmRegistryResponse } from "../types/NpmRegistry";
+import { DependencyEntityBuilder } from "@voodoogq/plugin-dependency-packages-node";
 
-export class NpmComponentEntityBuilder {
-  private builtEntities: ComponentEntity[] = [];
+export class NpmComponentEntityBuilder extends DependencyEntityBuilder {
+  public readonly dependencyTypeId: string = 'npm';
 
-  private template: EntityTemplate<ComponentEntity> = {
-    apiVersion: 'backstage.io/v1alpha1',
-    kind: 'Component',
-    metadata: {
-      packageVersionReferences: [],
-    },
-    spec: {
-      type: 'packageDependency',
-      // TODO: Need a way to set a default group
-      owner: 'guests',
-      // TODO: Need a way to set a default lifecycle
-      lifecycle: 'external',
-      dependencyOf: []
-    },
-  };
+  protected builtEntities: ComponentEntity[] = [];
 
   private registryMetaRetriever: RegistryMetaRetriever;
 
   constructor() {
+    super();
     this.builtEntities = [];
     this.registryMetaRetriever = new RegistryMetaRetriever();
-  }
-
-  // TODO: need to do some more name massaging, needs to be below 63 characters
-  private nameMutation(name: string): string {
-    const packageName = name.replace('@', 'at_').replace('/', '-');
-    return `dep.npm.${packageName}`
   }
 
   private npmRegistryUrl(name: string): string {
@@ -50,39 +32,6 @@ export class NpmComponentEntityBuilder {
     }
 
     return this.gitToHttpUrl(sourceLocation.url);
-  }
-
-  private gitToHttpUrl(gitUrl: string, path?: string): string {
-    // 1. Strip the git+ prefix
-    let url = gitUrl.replace(/^git\+/, '');
-
-    // 2. Extract branch if present
-    const hashIndex = url.indexOf('#');
-    let branch = '';
-    if (hashIndex >= 0) {
-      branch = url.slice(hashIndex + 1);
-      url = url.slice(0, hashIndex);
-    }
-
-    // 3. Strip .git suffix
-    url = url.replace(/\.git$/, '');
-
-    // 4. Build final URL
-    if (url.includes('github.com')) {
-      if (branch) {
-        url = `${url}/tree/${branch}`;
-      }
-      if (path) {
-        url = `${url}/${path.replace(/^\/+/, '')}`;
-      }
-    } else {
-      // Other platforms can be added here if needed
-      if (path) {
-        url = `${url}/${path.replace(/^\/+/, '')}`;
-      }
-    }
-
-    return url;
   }
 
   private compileLinks(meta: NpmRegistryResponse): { url: string, title: string }[] {
