+++ plugins/dependency-packages-node/package.json
@@ -27,6 +27,7 @@
   },
   "dependencies": {
     "@backstage/backend-plugin-api": "^1.3.1",
+    "@backstage/catalog-model": "^1.7.4",
     "@backstage/config": "^1.1.4",
     "@backstage/plugin-permission-common": "^0.8.4",
     "@backstage/types": "^1.2.1",
