+++ plugins/dependency-packages-node/src/builder/index.ts
@@ -1,0 +1,1 @@
+export { DependencyEntityBuilder } from './DependencyEntityBuilder';
