+++ plugins/dependency-packages-node/src/index.ts
@@ -1,2 +1,3 @@
+export * from './builder';
 export * from './dependencyTypes';
 export * from './extensionPoints';
